feat: move protocol contextualizers to dir
