+++ src/protocol/superchainFaucet/index.ts
@@ -1,0 +1,1 @@
+export { superchainFaucetContextualizer } from './superchainFaucet';
