+++ src/protocol/weth/index.ts
@@ -1,0 +1,1 @@
+export { wethContextualizer } from './weth';
